requirements.txt
torch>=2.2
transformers>=4.50,<6
datasets>=3.0
evaluate>=0.4
sacrebleu>=2.5
rouge-score>=0.1.2
bert-score>=0.3.13
accelerate>=1.0
sentencepiece>=0.2.0
PyYAML>=6.0
pandas>=2.2
numpy>=1.26
tqdm>=4.66
nltk>=3.9
